kisensum/openadr/requirements.txt
amqp==2.2.2
asn1crypto==0.23.0
attrs==17.3.0
behave==1.2.5
behave-django==0.5.0
billiard==3.5.0.3
bootstrap-admin==0.3.7.1
celery==4.1.0
certifi==2017.7.27.1
cffi==1.11.2
click==6.7
defusedxml==0.5.0
Django==1.11.5
django-activeurl==0.1.10
django-admin-interface==0.5.8
django-angular==1.1.3
django-appconf==1.0.2
django-bootstrap-datetimepicker==1.5
django-bootstrap-form==3.3
django-bootstrap-forms==0.1
django-bootstrap3==9.0.0
django-bootstrap3-datetimepicker==2.2.3
django-bootstrap3-datetimepicker-2==2.5.0
django-bootstrap4==0.0.3
django-classy-tags==0.8.0
django-colorfield==0.1.14
django-data-exports==0.7
django-datatable==0.3.1
django-datatables-view==1.14.0
django-datetime-widget==0.9.3
django-datetimepicker==3.14
django-debug-toolbar==1.8
django-filter==1.0.4
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-forms-bootstrap==3.1.0
django-inspect-model==0.7
django-tables2==1.11.0
django-widget-tweaks==1.4.1
djangorestframework==3.7.1
djangorestframework-xml==1.3.0
eight==0.4.2
et-xmlfile==1.0.1
factory-boy==2.9.2
Faker==0.8.6
future==0.16.0
generateDS==2.28.2
gevent==1.2.2
greenlet==0.4.12
idna==2.6
iso8601==0.1.12
isodate==0.6.0
jdcal==1.3
Jinja2==2.9.6
kombu==4.1.0
livereload==2.5.1
lxml==3.8.0
Markdown==2.6.9
MarkupSafe==1.0
mkdocs==0.17.1
names==0.3.0
numpy==1.13.3
odfpy==1.3.5
openpyxl==2.4.8
pandas==0.20.3
parse==1.8.2
parse-type==0.4.2
pluggy==0.6.0
psycopg2==2.7.3.1
py==1.5.2
pycparser==2.18
pyOpenSSL==17.3.0
pytest==3.3.0
pytest-django==3.1.2
python-dateutil==2.6.1
pytz==2017.2
PyYAML==3.12
selenium==3.6.0
signxml==2.4.0
six==1.11.0
sqlparse==0.2.4
tablib==0.12.1
text-unidecode==1.0
tornado==4.5.2
unicodecsv==0.14.1
vine==1.1.4
xlrd==1.1.0
xlwt==1.3.0
